Correctly remove expired jwts instead of sending them
